And all that the Lorax left here in this mess was a small pile of rocks with the one word, "unless." Whatever THAT meant, well, I just couldn't guess. That was long, long ago, and each day since that day, I've worried and worried and worried away. Through the years as my buildings have fallen apart, I've worried about it with all of my heart.
"BUT," says the Oncler, "now that you're here,
the word of the Lorax seems perfectly clear!
UNLESS someone like you cares a whole awful lot,
nothing is going to get better - it's not.
So... CATCH!" cries the Oncler.  He lets something fall.
"It's a truffula seed.  It's the last one of all!

"You're in charge of the last of the truffula seeds.
And truffula trees are what everyone needs.
Plant a new truffula -- treat it with care.
Give it clean water and feed it fresh air.
Grow a forest -- protect it from axes that hack.
Then the Lorax and all of his friends may come back!"

--------------------------------------------
Wonderful messages brought to you by fortune. This commit was generated
by garden-gnome for _ _ _ _ _ and giggles <3.