My My, hey hey Rock and roll is here to stay	The king is gone but he's not forgotten It's better to burn out		This is the story of a Johnny Rotten Than to fade away		It's better to burn out than it is to rust My my, hey hey			The king is gone but he's not forgotten
It's out of the blue and into the black		Hey hey, my my
They give you this, but you pay for that	Rock and roll can never die
And once you're gone you can never come back	There's more to the picture
When you're out of the blue			Than meets the eye
And into the black
		-- Neil Young
		"My My, Hey Hey (Out of the Blue), Rust Never Sleeps"

--------------------------------------------
Wonderful messages brought to you by fortune. This commit was generated
by garden-gnome for _ _ _ _ _ and giggles <3.